recall/bert_like/kuaiformer/practice/requirements.txt
tensorflow==2.12.0
numpy==1.24.3
faiss-cpu==1.7.4
scikit-learn==1.3.0
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2